+++ apps/prairielearn/src/components/AssessmentBadge.tsx
