+++ apps/prairielearn/src/components/AssessmentScorePanel.tsx
