+++ apps/prairielearn/src/components/EditQuestionPointsScore.tsx
