+++ apps/prairielearn/src/components/GroupWorkInfoContainer.tsx
