+++ apps/prairielearn/src/components/InstructorInfoPanel.tsx
