+++ apps/prairielearn/src/components/QuestionScore.tsx
